+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/emailPro/pack.emailPro.constant.js
@@ -1,0 +1,3 @@
+angular.module('managerApp').constant('PACK_EMAILPRO', {
+  webmailUrl: 'https://www.ovh.com/fr/mail/',
+});