+++ packages/manager/apps/telecom/src/app/telecom/telephony/alias/configuration/agents/ovhPabx/telecom-telephony-alias-configuration-agents-ovhPabx-modal.component.js
@@ -1,0 +1,13 @@
+angular.module('managerApp').controller('telecomTelephonyAliasConfigurationAgentsOvhPabxModal', class telecomTelephonyAliasConfigurationAgentsOvhPabxModal {
+  constructor($uibModalInstance) {
+    this.$uibModalInstance = $uibModalInstance;
+  }
+
+  cancel(message) {
+    return this.$uibModalInstance.dismiss(message);
+  }
+
+  close() {
+    return this.$uibModalInstance.close(true);
+  }
+});