+++ packages/manager/apps/telecom/src/app/telecom/telephony/alias/configuration/queues/ovhPabx/telecom-telephony-alias-configuration-queues-ovhPabx.constants.js
@@ -1,0 +1,5 @@
+export const NUMBER_EXTERNAL_TYPE = 'external';
+
+export default {
+  NUMBER_EXTERNAL_TYPE,
+};