+++ packages/manager/apps/telecom/src/app/telecom/telephony/alias/portability/order/telecom-telephony-alias-portability-order.js
@@ -1,0 +1,13 @@
+angular.module('managerApp').config(($stateProvider) => {
+  $stateProvider.state('telecom.telephony.portabilityOrder', {
+    url: '/portabilityOrder',
+    views: {
+      'telephonyView@telecom.telephony': {
+        templateUrl: 'app/telecom/telephony/alias/portability/order/telecom-telephony-alias-portability-order.html',
+        controller: 'TelecomTelephonyAliasPortabilityOrderCtrl',
+        controllerAs: 'PortabilityOrderCtrl',
+      },
+    },
+    translations: { value: ['.'], format: 'json' },
+  });
+});