+++ packages/manager/apps/telecom/src/app/telecom/telephony/billingAccount/billing/bill/telecom-telephony-billing-account-billing-bill.js
@@ -1,0 +1,13 @@
+angular.module('managerApp').config(($stateProvider) => {
+  $stateProvider.state('telecom.telephony.billing.bill', {
+    url: '/bill',
+    views: {
+      'telephonyView@telecom.telephony': {
+        templateUrl: 'app/telecom/telephony/billingAccount/billing/bill/telecom-telephony-billing-account-billing-bill.html',
+        controller: 'TelecomTelephonyBillingAccountBillingBillCtrl',
+        controllerAs: 'BillingAccountBillCtrl',
+      },
+    },
+    translations: { value: ['.'], format: 'json' },
+  });
+});