+++ packages/manager/apps/telecom/src/app/telecom/telephony/fax/assist/orders/telecom-telephony-fax-assist-orders.js
@@ -1,0 +1,13 @@
+angular.module('managerApp').config(($stateProvider) => {
+  $stateProvider.state('telecom.telephony.fax.assist.orders', {
+    url: '/orders',
+    views: {
+      'faxView@telecom.telephony.fax': {
+        templateUrl: 'app/telecom/telephony/service/assist/orders/telecom-telephony-service-assist-orders.html',
+        controller: 'TelecomTelephonyServiceAssistOrdersCtrl',
+        controllerAs: 'OrdersCtrl',
+      },
+    },
+    translations: { value: ['../../../service/assist/orders'], format: 'json' },
+  });
+});