+++ packages/manager/apps/telecom/src/components/telecom/telephony/alias/members/telecom-telephony-alias-members-add-modal.component.js
@@ -1,0 +1,13 @@
+angular.module('managerApp').controller('telecomTelephonyAliasMembersAddModal', class telecomTelephonyAliasMembersAddModal {
+  constructor($uibModalInstance) {
+    this.$uibModalInstance = $uibModalInstance;
+  }
+
+  cancel(message) {
+    return this.$uibModalInstance.dismiss(message);
+  }
+
+  close() {
+    return this.$uibModalInstance.close(true);
+  }
+});