+++ packages/manager/apps/telecom/src/components/telecom/telephony/scheduler/filters/telephony-scheduler-filters.component.js
@@ -1,0 +1,13 @@
+(function () {
+  angular.module('managerApp').component('telephonySchedulerFilters', {
+    require: {
+      telephonySchedulerCtrl: '^telephonyScheduler',
+    },
+    templateUrl: 'components/telecom/telephony/scheduler/filters/telephony-scheduler-filters.html',
+    controller: 'TelephonySchedulerFiltersCtrl',
+    bindings: {
+      filters: '=ngModel',
+      onChange: '&telephonySchedulerFiltersOnChange',
+    },
+  });
+}());