+++ packages/manager/apps/telecom/src/components/telecom/telephony/scheduler/params/telephony-scheduler-params.component.js
@@ -1,0 +1,9 @@
+(function () {
+  angular.module('managerApp').component('telephonySchedulerParams', {
+    require: {
+      telephonySchedulerCtrl: '^telephonyScheduler',
+    },
+    templateUrl: 'components/telecom/telephony/scheduler/params/telephony-scheduler-params.html',
+    controller: 'TelephonySchedulerParamsCtrl',
+  });
+}());